requirements.txt
streamlit
sentence-transformers
faiss-cpu
python-docx
python-pptx
pymupdf
pandas
openai==0.28